Moved queries and validation folders to be alongside the controllers folder & edited camelCase server files to be kebab-case Relates #69
