+++ server/controllers/auth/index.ts
@@ -1,4 +1,4 @@
 export { default as signup } from './signup';
 export { default as login } from './login';
-export { default as logOut } from './logout';
+export { default as logout } from './logout';
 export { default as getUser } from './get-user';
