+++ server/controllers/auth/logout.ts
@@ -1,7 +1,5 @@
 import { Request, Response } from 'express';
 
-const logOut = (req:Request, res:Response) => {
+export default (req: Request, res: Response) => {
   res.clearCookie('token').json({ message: 'Logged Out Successfully' });
 };
-
-export default logOut;
