+++ server/controllers/student/index.ts
@@ -1,6 +1,6 @@
-export { default as leaderboard } from './getLeaderboard';
+export { default as leaderboard } from './get-leaderboard';
 export { default as addPrivateQuizScore } from './add-private-quiz-score';
 export { default as sendEmail } from './send-email';
 export { default as checkUserAttendQuiz } from './check-user-attend-quiz';
-export { default as getQuiz } from './getQuiz';
+export { default as getQuiz } from './get-quiz';
 export { default as getQuestions } from './get-questions';
