+++ server/controllers/teacher/get-enrolled-students.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Response } from 'express';
-import { getEnrolledStudentsQuery } from '../../database/queries';
+import { getEnrolledStudentsQuery } from '../../queries';
 import { UserAuth } from '../../interfaces';
 
 export default async ({ params: { quizId }, user }:UserAuth, res:Response, next:NextFunction) => {
