--- server/database/queries/student/index.ts
@@ -1,7 +1,0 @@
-export { default as getLeaderboardQuery } from './getLeaderboardQuery';
-export { default as addPrivateQuizScoreQuery } from './add-private-quiz-score-query';
-export { default as getQuizDetailsQuery } from './get-quiz-details-query';
-export { default as checkUserAttendQuizQuery } from './check-user-attend-quiz-query';
-export { default as getStudentEmailQuery } from './get-student-email-query';
-export { default as getQuizQuery } from './getQuiz';
-export { default as getQuestionsQuery } from './get-questions-query';
