--- server/database/queries/teacher/index.ts
@@ -1,6 +1,0 @@
-export { default as getEnrolledStudentsQuery } from './enrolled-students';
-export { default as deleteQuiz } from './delete-quiz';
-export { default as getMyQuizzes } from './get-my-quizzes';
-export { createQuizQuery, createQuestionQuery } from './create-new-quiz-queries';
-export { default as userInfo } from './get-profile';
-export { default as checkQuizExistsQuery } from './check-quiz-exists';
