+++ server/queries/auth/check-email-taken.ts
@@ -1,6 +1,6 @@
 import escape from 'pg-escape';
-import { Checkable } from '../../../interfaces';
-import dbConnection from '../../config/connections';
+import { Checkable } from '../../interfaces';
+import dbConnection from '../../database/connections';
 
 export default ({ destination, email }: Checkable) => dbConnection.query(
   escape('SELECT * from %I WHERE email = $1', destination),
