+++ server/queries/auth/index.ts
@@ -1,2 +1,2 @@
-export { default as createNewUserQuery } from './signup-queries';
+export { default as createNewUserQuery } from './signup';
 export { default as checkEmailTakenQuery } from './check-email-taken';
