+++ server/queries/student/get-student-email.ts
@@ -1,4 +1,4 @@
-import dbConnection from '../../config/connections';
+import dbConnection from '../../database/connections';
 
 export default (userId: number) => dbConnection.query(
   `
