+++ server/queries/student/index.ts
@@ -1,0 +1,7 @@
+export { default as getLeaderboardQuery } from './get-leaderboard';
+export { default as addPrivateQuizScoreQuery } from './add-private-quiz-score';
+export { default as getQuizDetailsQuery } from './get-quiz-details';
+export { default as checkUserAttendQuizQuery } from './check-user-attend-quiz';
+export { default as getStudentEmailQuery } from './get-student-email';
+export { default as getQuizQuery } from './get-quiz';
+export { default as getQuestionsQuery } from './get-questions';
