+++ server/queries/teacher/check-quiz-exists.ts
@@ -1,4 +1,4 @@
-import dbConnection from '../../config/connections';
+import dbConnection from '../../database/connections';
 
 export default (quizId: string) => dbConnection.query(
   `
