+++ server/queries/teacher/delete-quiz.ts
@@ -1,4 +1,4 @@
-import connections from '../../config/connections';
+import connections from '../../database/connections';
 
 export default (quizId: String, teacherId: Number) => connections
   .query('DELETE FROM quizzes WHERE id=$1 AND teacher_id = $2', [quizId, teacherId]);
