+++ server/queries/teacher/get-teacher-quizzes.ts
@@ -1,4 +1,4 @@
-import connections from '../../config/connections';
+import connections from '../../database/connections';
 
 export default (teacherId:Number) => connections
   .query(
