+++ server/queries/teacher/index.ts
@@ -1,0 +1,6 @@
+export { default as getEnrolledStudentsQuery } from './get-enrolled-students';
+export { default as deleteQuiz } from './delete-quiz';
+export { default as getMyQuizzes } from './get-teacher-quizzes';
+export { createQuizQuery, createQuestionQuery } from './create-new-quiz';
+export { default as userInfo } from './get-profile';
+export { default as checkQuizExistsQuery } from './check-quiz-exists';
