+++ server/routes/student/index.ts
@@ -9,10 +9,10 @@ import { checkAuth } from '../../middlewares';
 
 const router = Router();
 
-router.get('/questions/:quizId', checkAuth('student'), getQuestions);
 router.route('/leaderboard/:quizTitle').get(leaderboard);
-
-router.post('/score', checkAuth('student'), checkUserAttendQuiz, addPrivateQuizScore, sendEmail);
 router.get('/quiz/:quizId', getQuiz);
+router.use(checkAuth('student'));
+router.get('/questions/:quizId', getQuestions);
+router.post('/score', checkUserAttendQuiz, addPrivateQuizScore, sendEmail);
 
 export default router;
