+++ server/routes/teacher/index.ts
@@ -6,11 +6,11 @@ import { checkAuth } from '../../middlewares';
 
 const router = Router();
 
-router.get('/quiz/:quizId/enrolled-students', checkAuth('teacher'), getEnrolledStudents);
-
-router.get('/profile', checkAuth('teacher'), getProfile);
-router.delete('/quiz/:quizId', checkAuth('teacher'), deleteQuiz);
-router.get('/quizzes', checkAuth('teacher'), getQuizzes);
-router.post('/quiz', checkAuth('teacher'), createQuiz);
+router.use(checkAuth('teacher'));
+router.get('/quiz/:quizId/enrolled-students', getEnrolledStudents);
+router.get('/profile', getProfile);
+router.get('/quizzes', getQuizzes);
+router.post('/quiz', createQuiz);
+router.delete('/quiz/:quizId', deleteQuiz);
 
 export default router;
