+++ server/validation/add-score-schema.ts
